+++ tests/testthat/test-tabulate.R
@@ -1,0 +1,7 @@
+context("tabulate")
+
+test_that("ypr_tabulate_yield", {
+  yield <- ypr_tabulate_yield(ypr_population())
+  expect_identical(check_tabulated_yield(yield), yield)
+  expect_identical(yield$Type, c("actual", "optimal"))
+})