+++ NEWS.md
@@ -1,3 +1,4 @@
+- Added `ypr_yield_tabulate()` to tabulate actual and optimal yields.
 - Added `ypr_sr()` to get stock recruitment parameters and `ypr_plot_sr()` to plot
 - Remove `Capture` (as covered by single value) and `TotalMortality` from schedule and added `Spawning`.
 - Switched yield parameters `harvest` and `biomass` to `FALSE` by default.
@@ -6,7 +7,6 @@
 - Added `ypr_inst2inter()` and `ypr_inter2instant()` to convert mortality rates.
 - Switched `nu` for `M`.
 - Added `Mb` to scale mortality by length and `BH` to distinguish between Beverton-Holt and Ricker curves and `es` for probability of a mature fish spawning and `Sm` for spawning mortality probability and `Sp` for spawning power and vulnerability power `Vp` and  slot limit non-complicance parameter `Nc` and scaling parameter `R0`.
-- 
 
 # ypr 0.0.1